Add in Route-model binding help to use slugs.
Because, what good are slugs if you can’t taste them.

Added SluggableRouter - which enables a search by if if the received
value is an integer > 0, or uses the slug as text otherwise.

Added a SluggableRoutingServiceProvider which exchanges the $router in
Laravel.

Added some additional methods to the Trait to avoid duplicating code to
seek Models by Id or Slug.

Added some Tests.